WT-5242 Minimize checkpoints pinned during backup (#5559)
* WT-5242 Minimize checkpoints pinned during backup

Checkpoints have time values generated from a per-session monotonic
clock.  Use these values to preserve only the checkpoints that may
be needed by backup.  At the end of each checkpoint, we save a
time value from the same clock.  When we open a backup cursor we
use this save value as a limit on checkpoint preservation.  All
checkpoints with times values less than or equal to that are kept
until the backup cursor is closed.

Add Python functional tests to verify behavior.

Update documentation to describe new interaction of backup and
checkpoints

Co-authored-by: Keith A. Smith <keith.smith@mongodb>