fix dhandle->session_inuse counter leak in __evict_walk
In __evict_walk, the dhandle->session_inuse counter is increased and decreased to walk the tree but in case of errors, the code jumps over the code that decreases the counter and basically leaves the dhandle as in-use so it never gets deleted during sweeps.

This patch adds a fix to decrease the counter even in case of errors.