WT-6118 Missing checkpoint in backup
Change how we track the time of the most recent checkpoint.  Instead
of saving it after checkpoints complete, use the time value we assign
when initializing a new checkpoint structure.  Rename the variable
where we store this to better reflect new usage.

Re-enable deletion of checkpoints during backup and the tests of that
functionality.