WT-6042 Changes to checkpoint04 to remove or debug failures on Windows.
This change removes the use of timestamps as they were not being updated
properly for checkpoints. It prints out via self.pr more information, in
particular, the value of all interesting statistics prior to checking
them so that if we fail we can see all of them in results.txt.