WT-6284 History store isn't visible to wt dump (#5695)
When we moved WT_CURSTD_IGNORE_TOMBSTONE from being a session flag to a cursor flag, we broke wt dump. wt dump uses a dump cursor which delegates calls to the cursor API to a "child" cursor. We need to set the flag on the child cursor for this to work.

We also modified the check in __wt_txn_read such that we will read the on-disk value's stop time pair even with WT_CURSTD_IGNORE_TOMBSTONE enabled if it is globally visible. This makes wt dump less useful so let's make an exception here.