WT-6349 Don't perform on disk key removal as part of checkpoint
On disk key removal needs to clear all the history store content
for the key and it might take quite a bit of time when there are
multiple keys that are needs to be removed. To avoid checkpoint
waiting for all these cleanup, better to do this as part of eviction.