WT-5769 Search history store can potentially walk the whole history store tree (2nd attempt) (#5753)
When we're searching history store, ignore tombstones and visibility and do the checking in the caller by inspecting the cursor's time window. This avoids performance issues where cursors that don't have any records visible to them will have to traverse the entire history store in a single cursor operation such as "search_near" or "next" just to return WT_NOTFOUND.

We've also temporarily disabled history store verification. This was previously not working due to a misuse of the WT_CURSTD_IGNORE_TOMBSTONE flag so we've kept it disabled to prevent excessive noise in format testing.