WT-6330 Perform RTS on the history store as a final step (#5756)
As the history store is checkpointed at the last step of checkpoint
process, during this time it is possible that newer updates are written
to the history from the data store tables by eviction process that is
not present in the data store. Due to this, it is possible that sometimes
history store may have newer updates that are not present in the datastore
checkpoint. If any backup/unclean shutdown uses this inconsistent
checkpoint it is possible that history store may have data newer than
the stable timestamp and can cause data visibility issues.

To avoid this problem, while recovering the database, perform
rollback to stable on the history store at the end.