WT-6366 Off-by-one overflow in block-modification bitmaps for incremental backup
Fix off-by-one error that caused overflow of allocated memory buffer
when setting the last bit in the bitmap used to track block changes
for incremental backup.

Allow block sizes as small as 4KB to make it easier to test and trigger
this type of bug.

Modify test/format and test/csuite/incr_backup to use the smaller
block sizes.