WT-6381 RTS to remove hs record that is appended to the update chain
During aborting of on-disk update on a data store while replacing it
with the history store update that have both equal start and stop
timestamps led to not removing the history store record can led to
the update present in both data store and history store. To hanlde
this problem, RTS must remove the history store update that is getting
appended to the data store.