WT-6395 Reorder the history store inserts if there exists an update with same timestamp
The counter field that is used as part of the history store key for uniqueness
will reset once the server is restarted. If there exists some updates in the
history store and data store with the same timestamp, the next reconciliation
moving the existing data store update to history store must be inserted after
the existsing history store record. Todo that get the counter value from the
existing record if there exists any record with same key and timestamp and use
the next counter for the new record.