WT-6416 RTS to skip traversing unnecessary internal pages
Instead of traversing all the internal pages to find out the leaf
pages that have newer updates to be aborted, skip traversing some
of the internal pages when they are not present in the memory and
their aggregated durable timestamp is less than stable timestamp.

If an internal page is not in-memory tells that the entire sub tree
of that internal page must not be present in-memory.