Add an init_merge field to the aggregate time window structures, set it based on whether the structure was initialized for merging or as a single time window.
Change the WT_TIME_AGGREGATE_IS_EMPTY macro to make checks based on how the structure was
initialized, either for merging or as a single time window. This fixes problems where the
WT_TIME_AGGREGATE_IS_EMPTY() would be used to check if the time window had default values
as opposed to values set up for merging.

When using the WT_TIME_AGGREGATE_IS_EMPTY and WT_TIME_AGGREGATE_IS_UPDATE macros, assert
the aggregate structure was initialized correctly.

Rename WT_TIME_AGGREGATE_INIT_MAX to WT_TIME_AGGREGATE_INIT_MERGE, it's not a max value,
it's a state.

Remove the WT_TIME_WINDOW_INIT_MAX macro; the only consumer was the variable-length column
store reconciliation, and I believe that use was incorrect, we're not aggregating a time
window, we're simply tracking the current and previous time windows.