WT-6395 Reorder the history store inserts if there exists an update with same timestamp (#5797)
* WT-6395 Reorder the history store inserts if there exists an update with the same timestamp

The counter field that is used as part of the history store key for uniqueness
will reset once the server is restarted. If there exist some updates in the
history store and data store with the same timestamp, the next reconciliation
moving the existing data store update to history store must be inserted after
the existing history store record. Todo that gets the counter value from the
existing record if there exists any record with the same key and timestamp and uses
the next counter for the new record.

* Remove hs_counter in the btree as always use 0 or max +1 for the same timestamp update as the counter value to make the counter as per key local instead of btree level.

* Follow __btcur_update only reuse the middle of pinned pages for inserts.

Co-authored-by: Michael Cahill <michael.cahill@mongodb.com>