WT-6388 Fix-up out-of-order updates in the history store (#5813)
This is intended to be an improvement to how we currently handle out-of-order content. There are two main changes here:

The first one is to correct out-of-order timestamps when inserting updates into the history store if they are out-of-order with respect to other content in the update list.

The other change is to "fixup" out-of-order content already in the history store to more closely follow the ordering that we'd get with an in-memory update list. When we're about to insert a new update into the history store, we check whether we have higher timestamps already in the history store for the same key. If so we "move" them back to the timestamp we're inserting by removing them and reinserting them just before inserting our current update.

To facilitate the second change, I've refactored the __hs_insert_record_with_btree_int helper to make it more general so that I can reuse it in the "fixup" logic.