WT-6448 RTS to read all the required pages as part of tree walk
RTS used to read internal pages as part of tree walk and read
the leaf pages internally and evict them is causing the tree walk
to be restarted. Instead of inventing the tree walk restart logic
again in RTS, better to load all the RTS required pages as part
of tree walk.