WT-6274 Leak of HS cursor and flags in eviction server
Change __wt_hs_cursor_close() to always restore prior value of session's
history store cursor so that it is symmetric with __wt_hs_cursor which
always sets those flags.

Also update comment in session close to make it clear that session->hs_cursor
doesn't need to be explicitly closed because it is incluced in the session's
list of open cursors.