WT-6448 RTS to read all the required pages as part of tree walk (#5841)
RTS used to read internal pages as part of tree walk and read
the leaf pages internally and evicts them is causing the tree walk
to be restarted. Instead of inventing the tree walk restart logic
again in RTS, better to load all the RTS required pages as part
of a tree walk.

The advantage of this approach is that eviction can run in parallel
with RTS on the same tree.