WT-6477 format.sh ignores failures if run directory never created (#5852)
* List the program name in the exit/abort message.

* Rework format.sh to handle missing RUNDIR directories, they may not exist if format fails
before the directory is created. If there's no RUNDIR directory, quit immediately as we
can't know if we've already reported a failure.

Rework format.sh's main loop to check on jobs as they're created rather than creating
as many as we can before checking. That way if there's a configuration problem we die
immediately rather than spinning off a bunch of jobs that aren't interesting.

Add the -R flag to the usage statement.

Don't use grep with awk, awk has its own pattern matching support.

* The monitoring program looks for the "process running" line in the log file, push it out quickly.

* Fix $(basename) call, don't discard the path.
There's a (small) window where we can get an invalid process ID, restore the check that handles it.