WT-6453 Pin transaction ids for history store cursor operations (#5858)
Since we're running all history store operations under read-uncommitted isolation, we don't need to guard the cursor with this flag to prevent it from acquiring a snapshot. We also want it to pin the last running transaction id like a normal read-uncommitted transaction would (which this flag is preventing it from doing) to avoid having updates freed out from underneath us.