WT-6508 Perform garbage collection on data store as part of checkpoint
To remove the obsolete pages that contains all the deleted keys,
the garbage collection as part of checkpoint must scan these pages
and remove them. This is used to do for the history store pages, now
it is enabled to perform on data store pages also.

To avoid reading unnecessary pages into cache, during tree walk,
the pages that doesn't have stop timestamp are ignored until they
are present in the cache.