WT-6521 Distinguish the difference between dirty and cleanup dirty
As part of checkpoint cleanup, the pages that cannot be fast deleted
are marked them dirty and let them to be reconciled in the next checkpoint
but this leads to problems for the operations that are expecting clean
trees. To avoid the problem, split the dirty flags into normal and cleaup.
Only raise errors when the tree is dirtied normaly and not for cleanup.