WT-6499 Ignore evict priority for btrees that are dominating cache usage. (#5905)
As we create more collections, the metadata file grows. In a system with lots (multiple thousands) of dhandles but barely any updates, the metadata file can dominate cache usage.

In the case where the metadata file causes the cache usage to go past any of the targets, the eviction server will kick in. But it will intentionally ignore the metadata file due to its evict priority until it exceeds the trigger and we begin evicting aggressively. In an idle system where cache usage isn't going up and we never begin evicting aggressively, the eviction server will continuously look for pages to evict without any success leading to high CPU usage.

My change is to ignore the evict priority if the file is currently contributing to half or more of the target cache usage. This seems like a sensible rule generally speaking and it does solve the high CPU usage we were seeing in this workload.