WT-6277 Compatibility tests verify failure in timestamp validation (#5811)
* Remove unnecessary braces.

* Add an init_merge field to the aggregate time window structures, set it based on whether
the structure was initialized for merging or as a single time window.

Change the WT_TIME_AGGREGATE_IS_EMPTY macro to make checks based on how the structure was
initialized, either for merging or as a single time window. This fixes problems where the
WT_TIME_AGGREGATE_IS_EMPTY() would be used to check if the time window had default values
as opposed to values set up for merging.

When using the WT_TIME_AGGREGATE_IS_EMPTY and WT_TIME_AGGREGATE_IS_UPDATE macros, assert
the aggregate structure was initialized correctly.

Rename WT_TIME_AGGREGATE_INIT_MAX to WT_TIME_AGGREGATE_INIT_MERGE, it's not a max value,
it's a state.

Remove the WT_TIME_WINDOW_INIT_MAX macro; the only consumer was the variable-length column
store reconciliation, and I believe that use was incorrect, we're not aggregating a time
window, we're simply tracking the current and previous time windows.

* If the parent aggregated time window is empty, use the current stable timestamp (as derived from
the set stable timestamp, or the recovery timestamp, if the stable timestamp has not yet been
set), as the base timestamp. This resolves an upgrade/downgrade problem in the 4.4 release: we
can't tell if the aggregated time window is just empty or if the internal pages were rewritten
by a 4.2 release and so are empty because time information wasn't maintained. We can't figure
out which is which, so in the case of an empty aggregated time window, we assert that any time
information we find is is before the current stable timestamp, which should be true in either
case.

* Cleanup pass, shuffle functions so we don't need forward declarations.

* Set the stable timestamp, otherwise the WiredTiger engine won't write a recovery timestamp. After
opening the database, query for the recovery timestamp and use it to initialize the timestamp value.

After re-opening the database, first verify the objects to catch portability problems sooner.

timestamp_parse() can be static to util.c.

Do minimal error checkng on the timestamp returned to us by the WiredTiger engine, it should be
fine, but make sure we can read it successfully.

* Only retrieve the recovery timestamp on the initial open of the database.