WT-6569 Squash the same transaction tombstone updates
Squash the same transaction updates whenever a tombstone is found
to write to disk to apply the stop timestamp on an earlier update
with different transaction. Incase if there is no update from
different transaction write them to disk to block the eviction.

While instantiating the prepared updates for tombstone, if the
update is also from the same transaction mark the update also
as prepare in-progress