WT-6578 Prevent reconciliation from looking past the on-disk value (#5958)
We've been seeing a similar pattern in a few recent build failures which are centered around concurrent eviction and checkpoint and more specifically, choosing updates for reconciliation that are earlier than the current on-disk value. This causes some updates to get inserted into the history store twice and can erroneously trigger out-of-order timestamp handling and lead to visibility issues.

This change involves reconciliation flagging each update that gets selected to be written to the data store. Subsequent reconciliations must select this update if they haven't already selected one regardless of visibility (i.e. they must not choose an older update than the one that was last selected). This should solve the problem of concurrent reconciliations undoing each others work.