WT-6708 Repurpose oldest start txn to be newest txn in the aggregated time window
The original intention of the oldest start transaction in the aggregated time windows
was to limit the search traversing into the page when none of the transactions are
visible to the current search snapshot. However the aggregated oldest start transaction
doesn’t include the history store values, due to which the original intention of this
aggregated oldest start transaction doesn't limit the search. So repurposing it to store
the newest transaction will have the other benefits.