Use the appropriate stretgth Arm barriers.
The existing dsb barriers guarantee completion, not ordering
and synchronize other system activities such as tlb and cache
maintenance instructions. Additionally, the sharability domain
of them is system. Both of these are stronger than required. A
dmb enforces ordering described in the header and the stronger
dsb isn't needed. Furthermore, given that all the entities in the
system t which are depending on the ordering are CPUs the sharbility
doesn't need to order accesses with respect to device or
non-cacheable memory, so it can be reduced to the inner-sharable
domain.

Update lint rules for the changes