WT-6708 Repurpose oldest start txn to be newest txn in the aggregated time window (#6027)
The original intention of the oldest start transaction in the aggregated time windows
was to limit the search traversing into the page when none of the transactions is
visible to the current search snapshot. However, the aggregated oldest start transaction
doesn’t include the history store values, due to which the original intention of this
aggregated oldest start transaction doesn't limit the search. So repurposing it to store
the newest transaction from both start and stop transactions present on the page will have
other benefits.