WT-6798 Utilize Arm LSE atomics and the correct strength barriers (#6042)
* Use Arm Large Scale Extensions (LSE) Atomics if available

outline-atomics preserves backward compatibility with Arm v8.0
systems but also supports using Arm v8.1 atomics. The latter can
massively improve performance on larger Arm systems. The flag was
backported to gcc7, 8, 9, and is the default in gcc10+.

If the compiler supports the flag use it.

bench/wtperf/runners/medium-btree.wtperf ops/s increases by 5x with
the flag applied

* Use the appropriate strength Arm barriers.

The existing dsb barriers guarantee completion, not ordering
and synchronize other system activities such as tlb and cache
maintenance instructions. Additionally, the sharability domain
of them is system. Both of these are stronger than required. A
dmb enforces ordering described in the header and the stronger
dsb isn't needed. Furthermore, given that all the entities in the
system t which are depending on the ordering are CPUs the shareability
doesn't need to order accesses with respect to device or
non-cacheable memory, so it can be reduced to the inner-shareable
domain.