RFC: Replace the yield instructions with an ISB
Use an isb instruction here to be closer to the original x86 pause instruction.
The yield instruction that was previously here is a nop that is intended to
provide a hint that a thread is a SMT systems could yield. This is different
from the x86 pause instruction which delays exeuction by O(100) cycles. The isb
will typically delay exeuction by about 50 cycles so it's a reasonable
alternative.

I've marked this with RFC beacuse while we've seen similar changes improve
performance on other projects, I don't know what the best way is to test a
high-contention senario that would result in the pausing being impactful here.
I'm looking for suggestions of tests that would illustrate the benefit.