WT-5691 Handle scenario where imported files have write generations ahead of current DB (#6067)
This PR is adjusting the write generation check (to figure out whether to wipe transaction IDs or not) to use a btree specific base write generation as opposed to `conn->base_write_gen`.

Since we want to keep the write gen ranges disjoint between runs (with the exception of the import case), we still want to keep `conn->base_write_gen` and all the machinery involved in keeping it in metadata, etc should remain as is.

The write generation for each btree is already written to metadata during checkpoint so we don't need to do anything special here.

The main changes necessary are:
- Keep track of what the `btree->write_gen` was on startup and store it in a `btree->base_write_gen`.
- Compare the page's write gen to `btree->base_write_gen` instead of `conn->base_write_gen`.

The test that I've added is a bit convoluted but it does demonstrate the difference between the new and old scheme. If I switch back to using `S2C(session)->base_write_gen`, then the test fails with an invisible record which is what we expect.