WT-6816 Design write gen scheme to allow dhandles with active history to get closed/re-opened (#6112)
As part of WT-5691, we switched to using a btree specific base_write_gen to determine whether or not to wipe transaction ids (as opposed to a connection-wide one which we used previously).

This could cause problems if we begin closing dhandles with active history since we'll wipe transaction ids . At the moment, we don't allow doing this with our dhandles but we intend to in the future so we should modify our write gen scheme to avoid this problem.

This is the second attempt at making this change. We had an issue where we'd open the history store btree as part of recovery before the connection-wide base write gen was properly initialized. As a result, the history store btree would have an incorrect base write gen and wipe the wrong transaction ids.