WT-6845 set session btree to HS when operating on history store
The session btree is accessed to find out the oldest_id and pinned_timestamp
specific to the btree during checkpoint is in progress to decide whether the
contents of the btree are obsolete or not. When operating on history store,
set the session btree to HS to get proper values to have a proper visibility.