WT-6845 Set session btree to HS when operating on history store (#6139)
The session btree is accessed during the calculations of oldest_id and pinned_timestamp when the checkpoint is in progress to decide whether the checkpoint transaction should be part of the calculation or not? The session btree doesn't point to HS when operating on history store leads to wrong calculation of the above values.