WT-6432 Add test case to misuse the timestamp API. (#6161)
Includes repeatable randomized calls to set 4 global timestamps, and within a transaction, set timestamps when transaction started, in the middle of transaction (before and/or after setting values), during prepare and at commit time.  Many of the calls are illegal, and the test verifies that no crashes or assertions arise.