WT-6568 Fix split generation use. (#6174)
* WT-6568 Fix split generation use.

Splits rely on a generation (aka epoch) to prevent data structures being
freed when they are reachable by a reader.  The pattern when overwriting
needs to be: (1) overwrite a data structure; (2) read the current
generation; (3) queue old data structures to be freed; (4) (optional)
switch to a new generation.

That is, we need to accept during splits that data structures cannot be
freed immediately.  Also, since other threads may be splitting in
parallel, the global split generation can be incremented at any time
(and new readers can start in the new generation).  So it is critical
that we read the generation after all new data has been published.
Subsequent readers (in at least that generation) will see the new data.

Incrementing the split generation could be deferred until a thread is
performing garbage collection.  However, since every eviction involves
some garbage collection and we don't want any thread to wait for all
readers to drain, this change bumps the generation after the split has
completed (and all overwritten data is queued to be freed in the current
generation).

* Add failing CONFIG to format directory.

* Add comment, refactor some gen code.

* wt_gen_next return void.

* Revert the refactoring of gen_oldest and wt_gen_active.

* Remove spurious comma.

Co-authored-by: Susan LoVerso <sue@mongodb.com>