WT-6743 Save and restore session's data handle during cursor reopen call (#6186)
Since the internal reopen call may be called on cursors unrelated to the current API call (as part of cursor cache sweeping), it needs to save and restore the session's data handle. This change also includes extra assertions.

Setting statistics for reopen now happens in the "file cursor" code, as that is where the session data handle is valid.