WT-7117 Apply only updates less than on-disk base update while restoring an update
Rollback to stable restores the full update from the history store modifies to replace
an on-disk data store update. During this process, it must apply only the history store
modifies that are less than the on-disk data store update, otherwise it can lead to
corrupted update that is being restored.