WT-6354 Added wt utility printlog tests
Extended the python util test set (test_utilX.py) with a series of
small tests for the 'printlog' functionality. This is intended to
extend the wt utility test coverage (where it was missing specific
test coverage for 'printlog')

To test printlog we assert the various output options of the
printlog utility and that they correspond with operations
performed on the given table.