WT-7089 Do not skip checkpointing objects with obsolete pages (#6231)
WiredTiger skips checkpointing objects that are not modified to speed
up the checkpoint process, but this can lead to problems to the objects
(like history store) that have obsolete pages that are needed to be
removed as part of checkpoint garbage collection.

To address the obsolete page cleanup from tables that are not modified,
check for any obsolete pages on the tables from a previous checkpoint
and do not skip them if there exists any.  Added a new btree flag to indicate
whether the btree has any obsolete pages. This flag is used to verify before
skipping the clean trees by the checkpoint.