WT-7117 RTS to skip modifies that are more than on-disk base update while restoring an update (#6261)
Rollback to stable restores the full update from the history store modifies to replace
an on-disk data store update. During this process, it must apply only the history store
modifies that are less than the on-disk data store update otherwise, it can lead to
a corrupted update that is being restored.

Also, during the search for a history store modify, use the restored update as a base
value instead of aborted on-disk value as the on-disk value is aborted by RTS, using
it as a base value can lead to wrong value returns.

Don't apply history store modifies that are aborted while forming a full update for the
selected update of the search.