WT-6354 Added wt utility downgrade tests
Extended the python util test set (test_utilX.py) with a series of
small tests for the 'downgrade' functionality. This is intended to
extend the wt utility test coverage (where it was missing specific
test coverage for 'downgrade').

To test downgrade we assert the log compatibility version is
accordingly changed when running `wt downgrade`.