WT-6354 Added wt utility upgrade tests
Extended the python util test set (test_utilX.py) with a series of
small tests for the 'upgrade' functionality. This is intended to
extend the wt utility test coverage (where it was missing specific
test coverage for 'upgrade')

To test upgrade we assert successful completion of the wt upgrade
function over a differing table data sets (by complexity).