WT-7159 Always write a full update to HS for on-disk updates
The updates written to the data store are always full updates, but
when these updates are moved into the history store by eviction due
to further updates on the key, make sure that on-disk update is always
written as full update even in the update list it is a modify, because
the on-disk update will be the base update for all the history store
updates that are previous to the on-disk update.

Due to concurrent operation of checkpoint and eviction, it is possible
that history store may have more recent versions of a key than the
on-disk version. Without a proper base value in the history store,
it can lead to wrong value being restored by the RTS.