WT-6354 Increase wt utility coverage with printlog, downgrade, upgrade tests (#6276)
* WT-6354 Added wt utility printlog tests

Extended the python util test set (test_utilX.py) with a series of
small tests for the 'printlog' functionality. This is intended to
extend the wt utility test coverage (where it was missing specific
test coverage for 'printlog')

* Added wt utility downgrade tests

Extended the python util test set (test_utilX.py) with a series of
small tests for the 'downgrade' functionality. This is intended to
extend the wt utility test coverage (where it was missing specific
test coverage for 'downgrade').

* Added wt utility upgrade tests

Extended the python util test set (test_utilX.py) with a series of
small tests for the 'upgrade' functionality. This is intended to
extend the wt utility test coverage (where it was missing specific
test coverage for 'upgrade')

* Printlog tests account for '-u' option
* Use session.log_printf in printlog test
* Structure comments with full sentences
* Generalize printlog LSN values
* Printlog test using LSN 1,0
* Global definitions for key and val string prefix
* Remove bitwise AND with 'self.print_user_data'