WT-7192 Fixes for cached cursor sweep. (#6309)
Simplify the WT_DHANDLE_CAN_REOPEN macro, it will no longer return true for some
weird flag combinations.  Use this macro when reopening as a final check.
When doing the sweep check, even if the cursor is otherwise
"reopenable", don't allow sweeping of a cursor that is using the current
dhandle. Remove the assertion that used to protect that case, as it
can no longer happen.