WT-6935 Update connection base write generation from all files of metadata
At the end of the recovery, the connection base write generation must be
updated with the latest checkpoint data from all the files. In some scenarios,
it is possible that the connection base write generation is already upto date
when no rollback to stable or roll forward occurs as part of recovery.