When a row-store page is read into memory, we find the largest group of prefix-compressed keys we can build from a previous key on the page plus the per-key suffix bytes, and save the start/stop information for the group in the WT_PAGE structure. It doesn't cost anything when the page is read into memory (it's just updating a few local variables in the main loop). It costs 8B in the WT_PAGE structure (80B to 88B), plus moving the page's read generation out of the "hot" chunk of the WT_PAGE structure.
When we look up a key on a leaf page, if the key is part of the page's group of prefix-compressed
keys, we unpack the key's on-page cell. If the key's prefix is entirely found in the group's
starting key, we copy the key's prefix and suffix into the caller's buffer and don't instantiate
anything in memory. The cost here is (1) unpacking an on-page cell and (2) doing two memory
copies into the caller's buffer every time we search for a key in the prefix group. Previously,
we built the key once and instantiated it in allocated memory on the first search, and subsequent
searches would point the caller's buffer at that instantiated memory instead of doing any cell
unpacking or memory copies. Obviously, this is a real cost, on second and subsequent searches
we're trading work to avoid the memory amplification. It wouldn't be difficult to detect a page
being repeatedly searched if we wanted to do that, and instantiating these keys in that case,
which would focus this change on point reads/updates vs. hot-cache workloads.

When we reconcile a page, we don't increase the key prefix compression if the previous key was also
prefix compressed and we don't gain at least 10 bytes over the previous key's compression. In
other words, if we had a key prefix, and the next key's prefix isn't at least 10B longer,
we use the same prefix for both keys. In my smoke tests, common strings plus appended record
numbers generate short groups of keys with a common prefix without this change, and with this
change they generate long groups of keys with a common prefix. Obviously, this is a real cost
as well, although the block compression of the page should pick up the common strings we are
choosing not to remove as part of prefix-compression.