WT-7150 Trailing un-init mem in schema project
When performing a schema project on a set of input columns we
grow our input cursors buffer with more memory than is required to
write out a given application columns. Whilst this doesn't impact
the overall functionality we do create trailing uninitialised
memory that will be flagged by clangs 'msan' as it gets reconciled.

The following changes adjust the buffer allocation calculations
to account for when the schema projection is re-using/overwriting
the memory of a previously allocated column.