Add a new configuration item, btree.prefix, which sets a common prefix on all keys of between 15 and 80 bytes. The btree.prefix configuration string also sets the btree.prefix_compression configuration, if it's not explicitly configured.
Fix a bug in general format configuration: the code assumed the library "prefix_compression"
default value was true, and it's actually false, which meant prefix compression is not being
tested in most cases.

Don't do any key generation work unless it's a row-store, that simplifies things. (I think it's
safe, but I'm not sure, column-store is currently broken so it's hard to test. I've added an
assert, so if I'm wrong it should be easy to find.)