Restructure in-memory key/value encoding to incorporate key prefix compression. The previous encoding supported 512B keys and 8KB values on 1MB pages; the new encoding supports 4KB keys plus key prefixes and 8KB values, but decreases the maximum supported page size to 128KB.
Change __wt_row_leaf_key_info() to return a key prefix when returning a physical key.

Change __wt_row_leaf_key_info() to not return a value. The calling convention is we get back a
reference to a cell, a reference to a WT_IKEY structure with an instantiated key plus a cell,
or a key, length and prefix triplet. This code is both ugly and confusing, but it needs to be
a fast code path.

Merge the functions that set the row-store WT_ROW reference when a page is read
(__wt_row_leaf_key_set() and __wt_row_leaf_key_set_cell()), into a single call by checking for
overflow keys in a common function rather than in the page-read function.

Change __wt_row_leaf_key_work() to handle on-page encodings that include prefixes. Where the
key can be immediately built from the page-wide prefix we calculated when the page was read,
do that. Otherwise, the prefix has to be rolled-forward or backward exactly like a prefix read
from a cell.

Change __cursor_row_slot_key_return() to handle prefix-encoded keys.

Change row-store leaf page reconciliation to handle prefix-encoded keys.

Minor changes along the way:

Change __cursor_row_slot_key_return() to remove unused variable kpack_used.

Change __wt_row_leaf_key_work() to not crack the key cell multiple times in the case of an
overflow key, it won't change under us, there's no reason.

Push the function to discard instantiated keys down into the btree_inline code to minimize
information in the upper layers of the btree code.

Don't initialize the time window in __wt_read_row_time_window() if it's going to be overwritten.
Rename __wt_row_leaf_value_exists() to __wt_row_leaf_value_is_encoded(), the test isn't if a
value exists, it's a fast path test for when the value is encoded and we aren't going to crack
the cell to find the time window.