Restructure the in-memory key/value encoding to make it possible to retrieve the on-page cell from all key encodings. The problem this solves is we have to be able to find the on-page key cell after instantiating a key (for example, to step past it to the value cell), which implies knowing the cell location when instantiating a key.  This wasn't previously a problem because only references to on-page cells were instantiated, that is, the cell location was always available during the instantiation step. Adding key prefixes into the key encoding scheme changed that: we might want to instantiate a key-prefix compressed key, and by encoding the key, we've given up the on-page cell location.  The trick is to always include a reference to the on-page cell in each encoding and store additional offsets to other pieces of the cell.
The change is wide-ranging because callers of the underlying key retrieval functions used the
fact of the cell being available in an encoding to indicate information about the encoding,
and this is no longer the case. There are additional changes to take advantage of earlier
changes that allow us to more quickly create full keys from prefix-compressed keys, based on
a page-wide, known prefix.